Services layers added
It is aimed to create a modular project structure suitable for layered architecture, in which domain services are used, by arranging business transactions to be carried out through services. Since data transfer between services and web applications will be via DTO models, DTO models will be created to be used in the return type and parameters of the relevant service methods.